refactor: use data from context instead of props
