+++ templates/types/streaming/nextjs/app/components/ui/chat/chat-message/chat-tools.tsx
@@ -1,15 +1,37 @@
+import { useChatMessage, useChatUI } from "@llamaindex/chat-ui";
+import { useMemo } from "react";
 import { ToolData } from "../index";
 import { Artifact, CodeArtifact } from "../widgets/Artifact";
 import { WeatherCard, WeatherData } from "../widgets/WeatherCard";
 
 // TODO: If needed, add displaying more tool outputs here
-export default function ChatTools({
-  data,
-  artifactVersion,
-}: {
-  data: ToolData;
-  artifactVersion?: number;
-}) {
+export default function ChatTools({ data }: { data: ToolData }) {
+  const { messages } = useChatUI();
+  const { message } = useChatMessage();
+
+  // build a map of message id to artifact version
+  const artifactVersionMap = useMemo(() => {
+    const map = new Map<string, number | undefined>();
+    let versionIndex = 1;
+    messages.forEach((m) => {
+      m.annotations?.forEach((annotation: any) => {
+        if (
+          typeof annotation === "object" &&
+          annotation != null &&
+          "type" in annotation &&
+          annotation.type === "tools"
+        ) {
+          const data = annotation.data as ToolData;
+          if (data?.toolCall?.name === "artifact") {
+            map.set(m.id, versionIndex);
+            versionIndex++;
+          }
+        }
+      });
+    });
+    return map;
+  }, [messages]);
+
   if (!data) return null;
   const { toolCall, toolOutput } = data;
 
@@ -31,7 +53,7 @@ export default function ChatTools({
       return (
         <Artifact
           artifact={toolOutput.output as CodeArtifact}
-          version={artifactVersion}
+          version={artifactVersionMap.get(message.id)}
         />
       );
     default:
